Intersect Exclusive Tests

+++ ProvingGround/src/UnitTest.cpp
@@ -229,7 +229,7 @@ std::vector<std::string> makeMacroString(const std::string& macroName,
 
 	for (const auto& val : macroDefTokens)
 	{
-		if (temp.size() + val.size() > 75ULL)
+		if (temp.size() + val.size() > static_cast<std::size_t>(75))
 		{
 			temp += " \\";
 			macroLines.emplace_back(std::move(temp));
